cleanup and optimizations

+++ base/plugins/iscore-plugin-scenario/Scenario/Document/Constraint/ConstraintHeader.hpp
@@ -35,7 +35,7 @@ public:
     m_view = view;
   }
 
-  static constexpr int headerHeight()
+  static constexpr double headerHeight()
   {
     return ConstraintHeaderHeight;
   }
